Can't have package-lock.json and yarn.lock
Heroku logs aren't complaining, but the build is still saying it has failed, which doesn't make sense to me.  Just figured I would delete this and then, worst case, we can install whatever is in here with Yarn.  Sophie's part looks to be fine with yarn instead of npm now, so this shouldn't be problematic regardless.