Updated How Search Works
Now hits google places API to get results.  Still need to integrate if a business already exists in the DB on that search, and to instead send that information to results.

Also modified how new Review works.  It gets the business ID that it is currently on so the review will have it for its record in the database.